chore: 디테일 관련 타입 이동
- #25
